# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: IncidentFlow MCP API
  description: Canonical API specification for IncidentFlow MCP. Generated from FastAPI routes and MCP
    tool metadata.
  version: 0.1.0
paths:
  /install.sh:
    get:
      tags:
      - ops
      summary: Installer script
      description: Return a curl-able installer script with a URL derived from current host.
      operationId: install_sh_install_sh_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '500':
          $ref: '#/components/responses/InternalServerError'
      security: []
...
  /.well-known/oauth-protected-resource:
    get:
      tags:
      - ops
      summary: OAuth protected resource metadata
      operationId: oauth_protected_resource__well_known_oauth_protected_resource_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '500':
          $ref: '#/components/responses/InternalServerError'
  /.well-known/oauth-protected-resource/mcp:
    get:
      tags:
      - ops
      summary: OAuth protected resource metadata (MCP path)
      operationId: oauth_protected_resource_mcp__well_known_oauth_protected_resource_mcp_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '500':
          $ref: '#/components/responses/InternalServerError'
  /.well-known/oauth-authorization-server:
    get:
      tags:
      - ops
      summary: OAuth authorization server metadata
      operationId: oauth_authorization_server_metadata__well_known_oauth_authorization_server_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '500':
          $ref: '#/components/responses/InternalServerError'
  /.well-known/openid-configuration:
    get:
      tags:
      - ops
      summary: OpenID Connect discovery metadata
      operationId: openid_configuration__well_known_openid_configuration_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '500':
          $ref: '#/components/responses/InternalServerError'
  /.well-known/jwks.json:
    get:
      tags:
      - ops
      summary: OAuth JWKS redirect
      operationId: oauth_jwks__well_known_jwks_json_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '500':
          $ref: '#/components/responses/InternalServerError'
  /oauth/register:
    post:
      tags:
      - ops
      summary: OAuth dynamic client registration bridge
      operationId: oauth_register_oauth_register_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '500':
          $ref: '#/components/responses/InternalServerError'
  /register:
    post:
      tags:
      - ops
      summary: OAuth dynamic client registration bridge
      operationId: oauth_register_register_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '500':
          $ref: '#/components/responses/InternalServerError'
  /authorize:
    get:
      tags:
      - ops
      summary: OAuth authorization redirect
      operationId: oauth_authorize_authorize_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '500':
          $ref: '#/components/responses/InternalServerError'
  /token:
    post:
      tags:
      - ops
      summary: OAuth token endpoint bridge
      operationId: oauth_token_token_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '500':
          $ref: '#/components/responses/InternalServerError'
  /revoke:
    post:
      tags:
      - ops
      summary: OAuth token revocation bridge
      operationId: oauth_revoke_revoke_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '500':
          $ref: '#/components/responses/InternalServerError'
  /.well-known/{challenge_path}:
    get:
      tags:
      - ops
      summary: OpenAI domain verification
      description: Return the configured OpenAI Apps domain-verification token.
      operationId: openai_domain_verification__well_known__challenge_path__get
      parameters:
      - name: challenge_path
        in: path
        required: true
        schema:
          type: string
          title: Challenge Path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    InternalServerError:
      type: object
      required:
      - detail
      properties:
        detail:
          type: string
          example: internal server error
  responses:
...
    InternalServerError:
      description: Unhandled exception
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternalServerError'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Bearer token auth. In development with no auth provider configured, /mcp may run unprotected.
...